organized project files, tidied up some code
