+++ app/src/androidTest/java/com/team33/qrcodepursuit/MainActivityTest.java
@@ -15,6 +15,8 @@ import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 
+import com.team33.qrcodepursuit.activities.MainActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -34,7 +36,7 @@ public class MainActivityTest {
     @Test
     public void changeFrag() {
         // in the future, change this to other fragments
-        onView(withId(R.id.bottomnavigation_menu_scan)).perform(click());
+        onView(withId(R.id.bottomnavigation_menu_home)).perform(click());
         onView(withId(R.id.container)).check(matches(isDisplayed()));
     }
