+++ app/src/main/res/layout/fragment_score_board.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScoreBoardFragment">
+    tools:context=".activities.Scoreboard.ScoreBoardFragment">
 
     <TextView
         android:id="@+id/title_text"
